Initial file upload
I was struggling to get the changes I made to establish the filter log and fix the 'get_brush_ID' function bug to upload to push to GitHub. I think it might have been because of the 300 SOW files that I deleted locally. So, deleted the old repository and created this new one.