Kennard-Stone subsetting in app
Added functionality to subset policies directly inside the app to whatever policies are in the global filter. This allows user to view most diverse K policies from the subset of policies remaining after filtering. Previously, Kennard Stone sampling was performed externally on all 463 policies and the data was loaded into the app.